Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/referrer-policy/origin-when-cross-origin/http-rp/same-origin/http-http/iframe-tag/swap-origin-redirect/same-origin-insecure.http-expected.txt



FAIL The referrer URL is origin when a
                                 document served over http requires an http
                                 sub-resource via iframe-tag using the http-rp
                                 delivery method with swap-origin-redirect and when
                                 the target request is same-origin. assert_equals: Reported Referrer URL is 'origin'. expected "http://localhost:8800/" but got "http://localhost:8800/referrer-policy/origin-when-cross-origin/http-rp/same-origin/http-http/iframe-tag/swap-origin-redirect/same-origin-insecure.http.html"
